tests for storage access implementation
